Allow libFraunhoferAAC to be shared (or static) as needed
By changing the cc_library_static to cc_library, it can be included as a
static_library or shared_library as needed.

In order to allow the IAMF decoder and the AAC decoder to share the
library, this CL makes it possible for the library to be included as
shared.

Bug: 396187214
Flag: android.media.swcodec.flags.iamf_aac_flac
Test: build
Change-Id: I734bd9aa50333c5d337a2f1e5108161ec855c7b9